+++ src/main/java/site/scalarstudios/scalarpower/block/ScalarPowerBlocks.java
@@ -18,6 +18,7 @@ import site.scalarstudios.scalarpower.machines.generator.barometric.BarometricGe
 import site.scalarstudios.scalarpower.machines.generator.culinary.CulinaryGeneratorBlock;
 import site.scalarstudios.scalarpower.machines.generator.entropy.EntropyGeneratorBlock;
 import site.scalarstudios.scalarpower.machines.generator.geothermal.GeothermalGeneratorBlock;
+import site.scalarstudios.scalarpower.machines.generator.watermill.WaterMillGeneratorBlock;
 import site.scalarstudios.scalarpower.machines.grinder.DoubleGrinderBlock;
 import site.scalarstudios.scalarpower.machines.grinder.GrinderBlock;
 import site.scalarstudios.scalarpower.machines.macerator.DoubleMaceratorBlock;
@@ -61,6 +62,10 @@ public class ScalarPowerBlocks {
 
     public static final DeferredBlock<GeothermalGeneratorBlock> GEOTHERMAL_GENERATOR = registerBlock("geothermal_generator",
             GeothermalGeneratorBlock::new,
+            properties -> properties.strength(4.0F, 4.0F).requiresCorrectToolForDrops().sound(SoundType.IRON).mapColor(MapColor.STONE));
+
+    public static final DeferredBlock<WaterMillGeneratorBlock> WATER_MILL_GENERATOR = registerBlock("water_mill_generator",
+            WaterMillGeneratorBlock::new,
             properties -> properties.strength(4.0F, 4.0F).requiresCorrectToolForDrops().sound(SoundType.IRON).mapColor(MapColor.STONE));
 
     /* Machines */