+++ Changelog.md
@@ -39,6 +39,7 @@
 - Added Ender Battery
 - Added Geothermal Generator
 - Added Redstone Clock
+- Added Water Mill Generator
 - Change Machine Frame Models + Texture to be Block instead of Isometric projection of a block
   - Also applied for Steel Machine Frame
 - Renamed Glass Fiber wires to Fiber Glass Wires
@@ -48,6 +49,8 @@
 - Battery Sidedness (I/O Faces)
 - Crafting Recipe Revamp
 - Finish Wiki
-- Fluid Stuff (Potential)
+- Freezer Machine (Water → Ice → Packed Ice → Blue Ice)
+- Infinite Water Source Block
+- Liquifier Machine (Cobble → Lava)
 - Texture Name and Model Cleanup
 - Tooltips